backend/api/permissions.py
from rest_framework import permissions


class CurrentUserOrAdminOrReadOnly(permissions.IsAuthenticated):
    def has_object_permission(self, request, view, obj):
        user = request.user
        if type(obj) == type(user) and obj == user:
            return True
        return request.method in permissions.SAFE_METHODS or user.is_staff
